pyproject.toml
[tool.poetry]
name = "dm-driver-fastapi"
version = "0.1.0"
description = ""
authors = ["Tom V <vossy01@gmail.com>"]
readme = "README.md"
packages = [{include = "dm_driver_fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
uvicorn = "^0.23.2"
fastapi-pagination = "^0.12.12"
pillow = "^10.1.0"
pytest = "^7.4.3"
httpx = "^0.25.1"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
alembic = "^1.12.1"
typer = "^0.9.0"
rich = "^13.6.0"
black = "^23.10.1"
ruff = "^0.1.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
